+++ kernel/frame_buffer_config.hpp
@@ -1,6 +1,5 @@
 #pragma once
 
-#include <stdint.h>
 
 enum PixelFormat {
   kPixelRGBResv8BitPerColor,
@@ -8,9 +7,9 @@ enum PixelFormat {
 };
 
 struct FrameBufferConfig {
-  uint8_t* frame_buffer;
-  uint32_t pixels_per_scan_line;
-  uint32_t horizontal_resolution;
-  uint32_t vertical_resolution;
+  int* frame_buffer;
+  int pixels_per_scan_line;
+  int horizontal_resolution;
+  int vertical_resolution;
   enum PixelFormat pixel_format;
 };
